Add quiz link in config and implement quiz sending
A new environment variable QUIZ was added to the config.py file to allow for a customizable quiz link. A new file, send_quiz.py, was also created that includes a method to send this quiz link to all users who have a telegram ID. Furthermore, the pricelist.py file was renamed to send_pricelist.py to improve clarity and consistency.